errors: Break apart messages and add fmt.Formatter
This changes the `causer` interface we use for error chains to require
`fmt.Formatter` and a second `writeMessage` method.

    writeMessage(w io.Writer, verb string)
    // verb is "%v" or "%+v"

The `writeMessage` method writes the message/context for just the
current error to a writer. Implementations may use verb to decide how
they print nested errors or other information.

That lets us implement a generic `fmt.Formatter` (formatCauser) for all
such errors, which prints the messages as `context: reason` as usual for
`%v`, and as a multi-line message in the following form for `%+v`.

    context:
    reason

Because `fmt.Formatter` handles `%v` correctly, this also switches over
existing `Error()` implementations to rely on that with `fmt.Sprint`.

Finally, to make the resulting errors friendlier in a multi-line
context, some of them were reworded to put the function name at the end
of the sentence instead of the middle. That is,

    function "foo".Bar (baz.go:42) cannot be provided
    // becomes,
    cannot provide function "foo".Bar (baz.go:42)
    // and with %+v
    cannot provide function "foo".Bar
    	baz.go:42